+++ packages/events/i18n/en.js
@@ -1,6 +1,0 @@
-
-i18n.map('en', {
-	btn_events: "Events",
-	label_eventsuser: "Places in program",
-	label_noevents: "No event in program"
-});