+++ packages/events/i18n/es.js
@@ -1,6 +1,0 @@
-
-i18n.map('es', {
-	btn_events: "Eventos",
-	label_eventsuser: "Lugar en prevision",
-	label_noevents: "Ningun evento en programa"
-});