+++ packages/events/i18n/it.js
@@ -1,6 +1,0 @@
-
-i18n.map('it', {
-	btn_events: "Eventi",
-	label_eventsuser: "Luoghi in programma",
-	label_noevents: "Nessun evento in programma"
-});